# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API",
        "description": "API untuk payment gateway simple dengan integrasi Midtrans. \r\n *     \r\n *     Features:\r\n *     - JWT Authentication\r\n *     - Shopping Cart Management\r\n *     - Order Processing\r\n *     - Payment via Midtrans\r\n *     - Order Reminder (Scheduled)",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Admin - Lihat semua produk",
                "description": "Get semua produk untuk admin (termasuk yang out of stock)",
                "operationId": "9058305dfe0377e6cf7c5253ed664401",
                "responses": {
                    "200": {
                        "description": "Success - Return all products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products fetched successfully for admin."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Gaming Laptop"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15000000
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "image_url": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Hanya admin yang bisa akses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Admin - Buat produk baru",
                "description": "Create produk baru dengan upload gambar (multipart/form-data)",
                "operationId": "e78a4cd5a8e47d96daa7a5a7a13d3e30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gaming Laptop"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "High-end gaming laptop with RTX 4090"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "images[]": {
                                        "description": "Upload multiple images (optional)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Gaming Laptop"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000000
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}